requirements.txt
langgraph>=0.2.0
litellm>=1.30.0
chromadb>=0.4.22
streamlit>=1.32.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
